requirements.txt
django
djangorestframework
markdown
django-filter
django-cors-headers
psycopg2-binary
python-dotenv
Unipath